Elide empty string variations
Historically Soong has treated a variation of "" differently than
no variation at all.  This is causing an impedance mismatch with
TranstionMutators, which skip creating new varaints when only a
single "" variation is requested.  Since no new variant is created,
any future dependency that requests the "" variation gets a missing
dependency error.

This patch converts the rest of Soong to match TransitionMutators.
Any variant created with a "" variation is identical to having
not created a new variant.  When adding dependencies, any requested
empty variation is ignored.  If IncomingTransition returns "" the
variation is ignored.

Bug: 319288033
Test: all soong tests
Flag: EXEMPT refactor
Change-Id: I868d7e3d3f348e3634a447302a666c86bf5b772d