Expand input variant matching in applyTransitions
applyTransitions finds an appropriate variant to use as the target of
a call to IncomingVariation.  It would previously look at the list of
input variants that were present when the transition mutator originally
ran, finding one that was a subset of the requested variant.

This original method falls apart when a transition mutator returned a
single "" variant when it originally ran, as this result could either
mean that the transition mutator was uninterested in the module (which
is the case for a huge number of modules, and needs be carefully
optimized), or it could mean that the mutator is relevant to the module
but happened to only need a single variant (for example a platform-only
module when running the sdk mutator).  A module that only has a single
"" variant may still need IncomingVariation run on it to rewrite the
incoming dependency to "", but we can't save the input variants of
every uninteresting module that has a "" variant.

Expand the input variant matching in applyTransitions to be more
resiliant against modules that only return a single "" variant for
a transition mutator.  Rather than use a subset match, which can
be incorrect when empty "" variants are not kept, expliclty track
which mutators have already run, and check for an identical set of
variants of only those mutators.  If no matching input variants are
found, also check against the current full list of modules to handle
the case where input variants were not saved.

Bug: 319288033
Test: transition_test.go
Flag: EXEMPT refactor
Change-Id: I8f45bc7d55bd294e59f8dd007d4799ddd3f41e35