Split bpmodify command into a library
Many Soong tests use an Android.bp file template with %s mixed in, and
then use fmt.Sprintf to customize it for each subtest.  This gets
awkward, especially when multiple %s slots are necessary in the same
template.

We already have a command line bpmodify tool that can modify an
Android.bp file.  Refactor it into a bpmodify library, which can
then be used in tests to customize the Android.bp snippet for
each subtest.

Flag: EXEMPT test-only
Test: bpmodify/bpmodify_test.go
Test: bpmodify/cmd/main_test.go
Change-Id: I0085a74ac44222e85a9ffca05161c146ab52ff9a