Use unique.Handle for DepSets
DepSets need to be serializable and deserializable, but they are graph
and it is important that after serializing a deserializing two DepSets
that pointed to the same transitive DepSet continue to point to the
same transitive DepSet. This also has to hold when deserializing some
DepSets while recreating others during incremental compilation.

To satisfy this requirement convert DepSets to a wrapper around a
unique.Handle, which will ensure there is only ever one copy of a
DepSet that has the same contents.

Bug: 375276086
Test: depset_test.go
Flag: EXEMPT refactor
Change-Id: I7cc72caa0050446403f5a089f0c7b7cd92ae6477