Add a UniqueList that can store a slice in the unique package
UniqueList is a workaround for Go limitation that slices are not
comparable and thus can't be used with unique.Make.  It interns
slices by storing them in an unrolled linked list, where each node
has a fixed size array, which are comparable and can be stored using
the unique package.

Bug: 375276086
Test: TestUniqueList
Flag: EXEMPT host only
Change-Id: I8e5b8f4d415a2bacb400793eaf107be31a49334b