Change blueprint.BuildParams' Optional to Default
Soong's android.BuildParams uses Default instead of Optional, so the
difference is confusing. Also, the grand majority of the time you want
things to not be built by default, so this switch makes it so you don't
have to remember to add Optional: true every time you use
blueprint.BuildParams.

Test: Presubmits
Change-Id: I7048af490a3b315ae2db2716de04c8cd81c822ce