Optimize parallelVisit with batching
parallelVisit cannot keep multiple CPUs saturated when calling very
short visitors, as the overhead of synchronizing and coordinating
the worker goroutines is much higher than the time spent in the visitor.
The normal solution for distributing small work items is to batch them,
but that was previously not done here due to the complexity introduced
by the "pause" feature: any visitor can request to pause until the
visitor on another module has finished.  If the module being waited on
or any of its dependencies are later in the same batch then the
worker would deadlock.

This patch introduces batching without introducing deadlocks when
pausing by returning any uncompleted work in the batch when a pause
is requested, allowing that work to be completed by another worker.

This reduces overall soong_build analysis time from 1m58s to 1m27s,
or 26%.  Benchmark_paralellVisit, which measures the per-module time
for a noop visitor, drops from 1480 to 84.14 ns/module.

Test: Test_parallelVisit
Test: TestWalkDeps
Change-Id: Ie674cbd7c11778e11865027e72b3d4ff9dbfb661