Move waitingCount decrement into workers
Any CPU cycles used by the orchestrator could delay sending work to
the workers, which could leave any other available CPUs unsaturated.
Since the decrement of waitingCount is atomic, move it into the workers,
and let the workers report back any modules that are now ready if
waitingCount is zero.

A future improvement could handle the newly ready modules in the
same worker.

Test: Test_parallelVisit
Change-Id: I2b7fb98d5b9f53c275495b54694021393e1b73a2