Move cycle check into parallelVisit
The dependency cycle check in updateDependencies is slow and runs after
many of the mutators.  parallelVisit already does a dependency cycle
check for any newly added forward dependency cycles, but it only does
it when it detects that it has deadlocked, which keeps it out of the
hot path.  Update the parallelVisit dependency cycle checker to also
handle pre-existing dependency cycles, and remove the cycle checker
in updateDependencies.

Test: Test_parallelVisit
Change-Id: I4b6c6aeda35d74b4e16edf2112d7d63b4cca872e