Reduce some calls to updateDependencies
updateDependencies is slow, as it has to recompute all forward and
reverse dependencies.  It only runs when Blueprint has detected a
dependency tree modification that it can't handle by directly modifying
forwardDeps and reverseDeps.

It used to run for newly added reverse dependencies so that it could
also detect dependency cycles, but that is no longer necessary now
that parallelVisit does the dependency cycle check.  Update the
logic to directly modify forwardDeps and reverseDeps instead.

It also ran for newly created modules, but newly created modules have
no dependencies so updateDependencies is unnecessary.

updateDependencies will still run for mutators that create new variants,
as that makes complex modifications to the dependency tree.

Test: all Soong tests pass
Change-Id: Id8a97cc9eeb11da707d9aae6ddb29efa6d3a25b0