Remove transitionMutatorInputVariants
Blueprint stores intermediate copies of of modules for use as
IncomingTransitionContext.Module() when adding dependencies after
transition mutators.  They are currently freed before
GenerateAndroidActions, but the next CL will move GenerateBuildActions
into the final mutator, allowing it to be coalesced into the final
mutators, and there is no longer an appropriate time to free them.

Test: no change to soong outputs
Change-Id: Ib37d7521ebee2e3b3b22a8d346d5339f2091d720