Store a pointer to *moduleInfo in Modules
Add a blueprint.ModuleBase struct that must be embedded into every
implementation of blueprint.Module in order to store a pointer to
the *moduleInfo.

Also make ModuleProxy store the *moduleInfo instead of the Module
and implement the info() method on it.  This allows all the methods
that operate on Modules to use info(), which means getWrappedModule
is no longer necessary, as ModuleProxy.info() will return the same
value as the corresponding Module.info().  That allows simplifying
android.ModuleProxy, as it can now just wrap blueprint.ModuleProxy,
and it no longer needs the android.getWrappedModule method.

Test: all soong tests pass
Change-Id: I7b1caef9a9085556e914de4743b3061a30184377