Remove c.moduleInfo
c.moduleInfo is used for two purposes.  It primarily stores a map from
Module to *moduleInfo to find the blueprint metadata related to a
Soong module.  That is no longer necessary now that all Modules
embed a pointer to a *moduleInfo.  The second use is to iterate over
all modules, but that can be done through c.moduleGroups via
c.iterateAllVariants().

Keeping c.moduleInfo updated is difficult, as it is a global map
that must be updated in a single goroutine.  Removing it allows
removing logic in runMutator to add new variants to it and in
cloneModules to remove old Modules and add the new clones.

Removing it will also simplify deleting all references to the
Module after GenerateBuildActions to allow it to be freed, as
it is no longer present in a global map.

Test: all soong tests pass
Change-Id: I90cce0488085ce865d7218f18adab34cbba4b125