Explicitly say we don't support select on defaults
The defaults property is just supposed to be a mechanical way to reduce
boilerplate. It's useful to be able to use defaults before the
configuration is decided, which means that we can't use a select()
on that configuration to decide the value of the defaults property.

Instead of implying we could add support, explicitly call out that
select on defaults is not supported.

Fixes: 397315538
Test: soong tests
Change-Id: I803de1d94eab273094c9b81866ac7b6629100efc