Reimplement UniqueList with weak pointers
UniqueList was implemented as an unrolled linked list to work around
the Go limitation that slices are not comparable and thereforce cannot
be stored in a unique.Handle.  Go 1.24 introduced weak pointers, which
means unique-like behavior can be implemented on top of the standard
library.

Reimplement UniqueList as a sync.Map (wrapped in blueprint/syncmap)
of weak pointers to slices, where the key is a maphash of the
elements of the slice.

Test: TestUniqueList
Change-Id: I99c49ad96088da571074618484616bbf9ec5328b