Verify gob_gen outputs are up to date
The go tools expect generated sources to be checked into the source
tree.  This is convenient for IDEs to be able to see the generated
code, but runs the risk of the generated files being out of date.

Add a -verify mode that compares the generated contents to the
file on disk, and run it as a validation action on all go sources.

Test: m nothing
Change-Id: If14103b754f17556917acb4fd907f95adbe9dffd