Optimize CalculateHash
Use a pool for hashers so that the hasher object, the ptrs map and the
mapStateCache can be reused.

Make int64buf an [8]byte so that it can be inlined into the struct.

Preallocate the ptrs map to reduce allocations for the providers
that have many pointers.  The highest measured number of pointers
in a single provider was 15220.

Test: hash_provider_test.go
Change-Id: I0a523bfcb722b015aa404f5f1098a82b9fe44cdc