Fix first unpack not respecting replace_instead_of_append
Normally when unpacking values in blueprint, all property struct
fields are uninitialized. But sometimes the module's factory method
initializes them. In this case, we need to respect
replace_instead_of_append instead of always using append mode. This
is particularly noticable when a boolean property is initialized to
true in the factory, as booleans are "appended" by the boolean OR,
so it was not possible to set them back to false.

Bug: 427467681
Test: m nothing --no-skip-soong-tests
Change-Id: I0e01097fe82122f04d7a3bf414c16e1d39a225b3