Fix dependency cycle checking
The dependency cycle checking results for existing dependencies of
a module were being overwritten with the results for newly added
dependencies, causing dependency cycle errors to be lost in some
cases and an unhelpful assertion to panic instead.  Switch to a
single combined iterator over existing and newly added dependencies.

Fixes: 433694465
Test: Test_parallelVisit/existing_cycle
Change-Id: I12c24849d8df88dd3d665d1cabc7ec6944eafdf5