Reduce allocations in gobtool.DecodeString
17% of all allocations in an incremental analysis build are from
gobtools.DecodeString.  A byte buffer is created to hold the bytes
of the string, and then a copy is made to convert it to an
immutable string.  Since the byte buffer is about to be released
use unsafe.String to reuse the bytes in the buffer as the string
bytes, saving an extra allocation and string copy on every call.

Test: builds
Change-Id: I31acbeef1c19f977ffff1d2767e731fd310660be