Reland: Hash duplicate pointers consistently
Hashing a pointer the first time it is seen does a complicated sequence
to store the current hash, reset the hasher, hash the contents of the
pointer, and then hashing the previous hash value on top of the current
hasher state.  When a pointer is seen the second time it only hashes the
cached pointer hash value.  This results in different hashes for the same
struct if it contains unique pointers to structs that have the same
contents vs. two pointers to the same struct.  Since decoding from the
incremental cache doesn't restore duplicate pointers (unless they are
part of DepSet), this can results in hash mismatches after restore.

Make the first and second visits of the same pointer affect the hash
in the same way by reseting the hasher again after calculating the
pointer hash, and then always hashing the old value and the pointer
value.

This patch relands Idd68c5f5b4fa509b12815be39d0dc2e60533e8e2 with
a fix to include the new proptools dependency in the Android.bp
file.

Test: TestHashingDuplicatePointers
Test: TestEncDec
Bug: 445716445
Change-Id: I5f31beae05b2a074798771d814d32d69c0d46243