Cache providers by hash
Modules may have the same provider value, especially once we move the
point of incrementality earlier to include providers set by mutators
that are then copied to other variants when they are split.

Store the list of provider hashes on each module, but use the hash
of the provider as the key in the provider database so multiple
modules can share the same stored value.

Use a map to ensure values are only written to the database once,
and a syncmap to cache them after reading from the database.

Reduces first analysis time from 188 seconds to 151 seconds, and
the provider database size from 18GB to 14GB.

Test: context_test.go
Change-Id: I8e695a68f97a4c2a0696dae4ed239f9a2c387c3c