Introduce coreModuleInfo in moduleGroup
In preparation for supporting on demand variants, this CL adds a new
property in moduleGroup to store the "core" module information. This
will store the information after all properties have been unpacked, and
defaults have been applied. A subset of moduleInfo properties will have
a deep copy. Others might be added later if required.

This "core" moduleInfo will serve as the starting point for creating the
on demand variants, but it should be a noop for now.

Bug: 448182009
Test: m nothing --no-skip-soong-tests
Change-Id: I10cced6734cba0c98ce366ac536d446ee62c1521