Include type in hash
Hashing the contents of structs with identical contents will produce
the same hash, even if the underlying type of the struct is different.
For example, hashing the zero values of struct { s []string } and
struct{ p []android.Path } will be identical.

Include the type of the value passed to CalculateHash in the output,
as well as the types in any interface fields.

Test: hash_provider_test.go, hash_type_test.go
Fixes: 445750906
Change-Id: Iccd533c4931a4f4d7c91a885bee9768e5ac86c8f