Add dependency from test singleton to modules
TestSingletonNotRestoreForSingletonChange needs to trigger a singleton
provider hash to change.  It currently does that by changing a global
variable to rename a singleton, which changes its output provider value.
Make the singleton depend on the modules and change its output to be
a list of all of the modules, and modify the Android.bp file to force
the singleton's output to change.

Test: TestSingletonNotRestoreForSingletonChange
Change-Id: I437f662759cdd278ed4f9fa5c4ee355f58509f90