Basic variants on demand - leaf node, single transition, no mutate
This is a basic/starting implementation of on demand variant creation.
It introduces a `SplitOnDemand` to transition mutator. These will not be
split immediately. If a subsequent mutator requests this, the variant
will be created on demand from moduleGroups' coreModuleInfo.
Specifically, AddDependency* will add the onDemandDep to
mctx.onDemandDeps, and runMutator will read this from the globalStateCh
channel to create the dependency edges.

There are still a lot of gaps that need to be addresed. Here are the
ones that come to mind, but please feel free to suggest more
1. Reapply mutations from previously completed mutators to bring it to
   the same state it would be in had the split happened upfront
2. On demand variant creating its own demand variant in the current
   mutator that calls AddDependency*
3. Same as (2), except one of the previously completed mutators creates
   on demand variants

I will be doing these in followup work.

Bug: 448182009
Test: go test . -race
Change-Id: Ibd15c75c852454f99a6f44cd1e54a43f5a75ed33