Dedup `rerunMutatorsOnVariantOnDemand` of on demand variants.
Previously, the mutators will be run from AddDependency* of rdep. Since
`createVariantOnDemand` returns a "new" clone of `coreModuleInfo`, this
meant that the rerun would happen proportional to the no of rdeps
requesting an on demand variant. This CL dedupes this by moving `rerunMutatorsOnVariantOnDemand` into the
visit function of the on demand variants.

The flow is
- AddDependency* from rdep will queue on demand variants for visit, and
  will pause till visit of the dependency is done
- Visit from the on demand variant will communicate with the main
  goroutine to determine if multiple on demand variants have been
requested. One of them will have the mutators run on them, and the rest
will refer to this on demand variant via `createdOnDemandReplaceWith`
- cloneLogicModule has been moved from createVariantOnDemand to
  rerunMutatorsOnVariantOnDemand. This saves the expensive clone
operation of coreModuleInfo per rdep
- rdep unpauses, and AddDependency will update module.directDeps to the
  on demand variant `createdOnDemandReplaceWith`.
- The orchestrator will add the on demand variant to moduleGroup at the
  end.

Bug: 448182009
Test: go test . -run TestOnDemandDependendcyVariantLeafNode -race
Change-Id: I6e66bb78796225f83fd6281e9aa7aaba7b5610fc