On demand variants of transitive deps
Applying earlier mutators to on demand variants might reveal additional
depencies that also need to be created on demand. This CL adds these
transitive on demand variants to the build graph.

Details
- `rerunMutatorsOnVariantOnDemand` has a new pauseFunc arg to queue the
  transitive on demand variants for `visit` in `runMutator`.
- `findVariant` and `supportsOnDemandVariant` has additional args to
  provide the current mutator index. This is necessary since for "Non-Far"
variant match in `supportsOnDemandVariant`, we need to know whether the
rdep is in pre transition or post transition.
- transition mutator will use a new requestedOnDemandVariant (type
  variationMap) to set the transition info of the on demand variant.

Bug: 448182009
Test: go test -run TestOnDemandDependencyVariantTransitiveDeps -race
Change-Id: I96a964a24a54462a438f033c639f69e342f1ebd7