Variants on demand for multiple transition mutators.
The preceding work for b/448182009 has focussed on a single transition
mutator. This CL covers variants that can be created by applying
multiple transition mutators.

The tricky part is recreating the dep's IncomingTransition context
inside `applyTransitions` of rdep. The latter finds the closest matching
variant from previous transitions, and uses that as the context for
IncomingTransition. In "variant on demand", a closes matching variant is
not guaranteed to exist. This CL recreates the IncomingTransition
context by running the completed mutators across a moving window.

It is unclear whether this is the most optimal way to create the
IncomingTransition context. This will be revisted later if necessary.

Bug: 448182009
Test: go test . -run TestOnDemandDependendcyVariantMultipleTransitions
-race

Change-Id: I25fd656784fb31da2b40acaef281716f7a85b811