Prevent preview of future state of on demand variants
For normal variants, each mutator group commences after the previous
mutator group has been run on all the variants. This is currently not
true for on-demand variants, which would run all the mutator groups on
dep, followed by run of all the mutator groups on rdep.

This CL adds a unit test to verify that on-demand variants do not get a
preview of the future state of their deps. It relies on ag/36615796,
which switched the return type of Add*Dependency to ModuleProxy, and the
existing provider framework which raises an exception if a
MutatorProvider is accessed from a future mutator.

Bug: 448182009
Test: go test . -run TestNoPreviewOfFutureStateOfOnDemandVariants -race
Change-Id: I23bd0acc0a23b4b8235abf932fe4d9e4386166de