Memoize transitive provider hashes
Using ctx.walkDeps to calculate transitive provider hashes on the
920 modules that need it ends up visiting 18871044 modules,
significantly slowing down incremental analysis.  Store the hash
of transitive dependencies on each module so that the next module
only has to visit its direct dependencies.

Also store the hash of all of the providers of the module, which
speeds up the non-transitive case as well.

Bug: 457275077
Test: incremental_test.go
Change-Id: I1d62d76fafc3420603bb65b7f391d07a75ef5c1a