Optimize hashing to use CustomHash more often
Make CalculateHash a generic function that takes a CustomHash,
which should make reflect.TypeOf a constant and call the type's
CustomHash method directly.

Add CalculateHashReflection that uses CalculateHash if the type
implements CustomHash, or falls back to reflection if it does not.

Make Hasher.CalculateHashReflection call CustomHash if it finds
an interface that implements CustomHash.

Convert some hashes of known types to wrap the types in a CustomHash.

Test: incremental_test.go
Change-Id: Idde92f7b3ea1e1445c3b08c01ceda338d9f69131