Set forwardDeps and reverseDeps for on-demand variants
on-demand variants are currently added to directDeps of the requesting
rdeps, but not to `forwardDeps` and `reverseDeps`. This would cause
future mutators (topdown or bottomup) to run on the on-demand variants
in an arbitrary order. This CL uses a `newReverseDeps` property to fix
the traversal order of the on-demand variants.

Bug: 448182009
Test: m nothing with ag/36952357 (switch rust transition mutators to
on-demand)

Change-Id: Ice1dce9b58921dafdf9877e3c9412083ff61b492