Rewrite supportsOnDemandVariant to allow combanitorial variation maps
The current implementation of supportsOnDemandVariant compares each
mutator key against the module's corresponding SplitOnDemand values. This
has the following issues
1. Assumes that each mutator axis is independent. This is not true for
   Android mutators like (os,arch) where the supported arch depends on
the os value. e.g. (linux,x86_64) and (windows,x86) could be supported
variants but not (linux,x86) and (windows,x86_64).
2. Requires SplitOnDemand to return the pre-transition values as well.
   e.g. rust modules dont have a link:shared variant, and use
IncomingTransition to switch all requests to link:"". When rust modules
are switched to on-demand, supportsOnDemandVariant will return false.

This CL replaces this with a heuristic based searchOnDemandVariant (to
keep search limited), followed by a call to applyTransitions to rerun
all the mutators till the completed transition mutator. At each stage,
the resolved variation will be compared against SplitOnDemand. If the
resolved variation is not found, findVariant will return nil.

Bug: 448182009
Test: m nothing with ag/36952357 (switch rust transition mutators to
on-demand)

Change-Id: Ibf161115ed6a774db4a58470fd2d20a2066db737