Threadsafe implementation of forward/reverseDeps of on-demand variants
ag/37289407 fixed the forward and reverseDeps references of the
on-demand variants, but did so in a non threadsafe manner. Specifically,
it modified `newReverseDeps` of the "final" on-demand variant (which
gets added to group.modules) directly from the Add.*Dependency context
of the requesing rdep.

This CL fixes it by moving this info to `newOnDemandReverseDeps` of each rdep.

Bug: 467056360
Test: go test . -race -count=100
Change-Id: I1d404729507f35b475e81116d3f52a71a5b58f11