Don't reset orderOnlyStringsCache
The incremental database is effectively append-only.  Resetting the
orderOnlyStringCache and refilling it from only the currently
referenced order-only strings can cause the order-only cache to
be missing an entry that is still referenced from the incremental
database.  If the module that references the order-only cache entry
is re-added it can reuse the entry in the incremental database,
resulting in a panic when it can't find the referenced order-only
cache entry.

Bug: 466891621
Test: build/soong/tests/build_action_caching_test.sh
Change-Id: I419979508e0b063464e3dfd99f0c7b0c31bb9122