Add more advanced syntax to static commands
This change lets us use a new Command2 field instead of Command
in blueprint.RuleParams. Command2 can be constructed out of strings
and HostTools. The HostTools will resolve to both strings and
dependencies. This allows us to refactor how host tool deps are
implemented without changing all the static rules. (except we have
to change them once now to this new syntax)

Currently, the data provided by a HostTool must be statically known
without analyzing other modules. This is because CommandDeps is not
actually a feature of ninja; we copy the CommandDeps from the rules
in the soong representation to the builds in the ninja representation.
On top of that, incremental analysis may mean that modules are not
re-analyzed, but we do need to re-analyze their build definitions when
the rules they depend on have changed CommandDeps. So essentially we
need the ability for modules to depend on rules, which then depend
on modules again. To avoid that complexity, instead always emit static
values for the HostTool's command and deps. The static value for the
deps will just be a hostTool-<name>-deps ninja phony, and when the
host tool module is analyzed it will create that phony.

If we do ever add the ability for modules to depend on rules, we
shouldn't need to rewrite all the static rules again thanks to this
syntax.

Bug: 448484009
Test: SOONG_ACTION_SANDBOXING=nsjail m out/soong/.intermediates/external/libtextclassifier/java/TextClassifierService/android_common/dex_depfile_verifier/TextClassifierService.jar.depfile_verifier
Change-Id: Ibdc0f037989043141351b4bd1ad15489594ab75e