Inter-variant dep edges between on-demand variants
Soong allows a module variant to depend on an earlier variant of itself.
This is implemented by looking at the order in module.group.modules.
This runs into issues if both variants are created on demand, since the
on-demand variants are added to module.group.modules at the end of
`runMutator`.

This CL defers the inter-variant dependency check of on-demand variants
to the point where the on-demand variants are added to
module.group.modules.

Test: m nothing --no-skip-soong-tests with ag/37428203
Bug: 448182009
Change-Id: I9c1e3fdc7d8fec0b23f70dc6a1794906bd8d7d83